pyproject.toml
[project]
name = "kedro_azureml_pipeline"
dynamic = ["version"]
description = "Kedro plugin with Azure ML Pipelines support"
readme = "README.md"
maintainers = [
    { name = "Guillaume Tauzin", email = "gtauzin@stateful-y.io" }
]
authors = [
    { name = "Guillaume Tauzin", email = "gtauzin@stateful-y.io" }
]
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "kedro>=1.0.0",
    "kedro-datasets>=1.0.0",
    "pydantic>=2.6.4",
    "cloudpickle>=3.0.0",
    "azure-ai-ml>=1.2.0",
    "pyarrow>=18.0.0",
]

[project.optional-dependencies]
mlflow = [
    "azureml-mlflow>=1.42.0",
    "mlflow>=2.0.0",
    "kedro-mlflow>=2.0.0",
]

[project.entry-points."kedro.project_commands"]
azureml = "kedro_azureml_pipeline.cli:commands"

[project.entry-points."kedro.hooks"]
azure_local_run_hook = "kedro_azureml_pipeline.hooks:azureml_local_run_hook"
mlflow_azureml_hook = "kedro_azureml_pipeline.mlflow_hook:mlflow_azureml_hook"

[dependency-groups]
dev = [
    {include-group = "tests"},
    {include-group = "lint"},
    {include-group = "docs"},
    {include-group = "fix"},
]
tests = [
    "pytest>=8.3",
    "pytest-cov>=7.0",
    "pytest-mock>=3.14",
    "pytest-xdist>=3.8",
    "covdefaults>=2.3",
    "hypothesis>=6.100",
    "numpy>=2.1.0",
    "pandas>=2.2.0",
]
lint = [
    "ruff>=0.8.0",
    "ty>=0.0.12",
]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-autorefs>=1.0",
    "mkdocs-material>=9.7",
    "mkdocstrings[python]>=0.26.0",
    "pymdown-extensions>=10.20.1",
]
fix = [
    "pre-commit-uv>=4.1",
]

[project.urls]
Homepage = "https://github.com/stateful-y/kedro-azureml-pipeline"
Documentation = "https://kedro-azureml-pipeline.readthedocs.io"
Repository = "https://github.com/stateful-y/kedro-azureml-pipeline"
"Bug Tracker" = "https://github.com/stateful-y/kedro-azureml-pipeline/issues"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kedro_azureml_pipeline"]

[tool.hatch.version]
source = "vcs"
fallback-version = "0.0.0"

[tool.hatch.build.hooks.vcs]
version-file = "src/kedro_azureml_pipeline/_version.py"

[tool.ruff]
line-length = 120
target-version = "py311"
src = ["src", "tests"]
show-fixes = true
exclude = ["src/kedro_azureml_pipeline/_version.py"]

[tool.ruff.format]
preview = true
docstring-code-format = true
docstring-code-line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "PL",  # Pylint
    "T201", # Print statement
]
ignore = [
    "E203",    # space before : (needed for how black formats slicing)
    "E501",    # line too long (handled by formatter)
    "B008",    # do not perform function calls in argument defaults
    "PLC0415", # import outside top-level (lazy imports used intentionally)
    "PLR0913", # too many args in func def
    "PLR0915", # too many statements
    "PLR0912", # too many branches
    "PLR2004", # magic value used in comparison
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["ARG", "S101", "T201", "PLR2004"]  # Allow assert statements, prints, and magic values in tests
"docs/**/*" = ["T201", "ARG001", "PLW0603", "SIM105", "SIM108"]  # Allow prints, unused args, globals, and simplifiable patterns in documentation hooks
"src/kedro_azureml_pipeline/cli.py" = ["ARG001"]  # Click callbacks receive unused context
"src/kedro_azureml_pipeline/cli_functions.py" = ["ARG001"]  # Click callbacks receive unused context/param
"src/kedro_azureml_pipeline/hooks.py" = ["ARG002"]  # Kedro hook signatures require specific args
"src/kedro_azureml_pipeline/mlflow_hook.py" = ["ARG002"]  # Kedro hook signatures require specific args
"src/kedro_azureml_pipeline/generator.py" = ["ARG002"]  # Method overrides require specific args

[tool.ruff.lint.isort]
known-first-party = ["kedro_azureml_pipeline"]

[tool.interrogate]
ignore-init-method = true
ignore-init-module = true
ignore-magic = false
ignore-semiprivate = false
ignore-private = false
ignore-property-decorators = false
ignore-module = false
ignore-nested-functions = false
ignore-nested-classes = true
ignore-setters = false
fail-under = 100
exclude = ["setup.py", "docs", "build", "tests", "*_version.py"]
ignore-regex = ["^get$", "^mock_.*", ".*BaseClass.*"]
verbose = 0
quiet = false
whitelist-regex = []
color = true
omit-covered-files = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--color=yes",
    "--cov=kedro_azureml_pipeline",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "integration: marks tests as integration tests that run subprocesses (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.coverage.paths]
source = ["src/kedro_azureml_pipeline"]

[tool.coverage.run]
source = ["kedro_azureml_pipeline"]
branch = true
parallel = true
sigterm = true
concurrency = ["multiprocessing"]
plugins = ["covdefaults"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/_version.py",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
fail_under = 1
include_namespace_packages = true

[tool.coverage.html]

[tool.rumdl]
flavor = "mkdocs"
line-length = 120
disable = [
    "MD012",  # Multiple consecutive blank lines (common in generated content)
    "MD013",  # Line length (handled by formatter, many long URLs/tables)
    "MD022",  # Heading blank lines (varies by document style)
    "MD031",  # No blank line before fenced code block
    "MD032",  # List should be preceded by blank line
    "MD033",  # Inline HTML (needed for README logo, MkDocs grid cards)
    "MD034",  # URL/email without angle brackets
    "MD045",  # Image missing alt text (logos use empty alt intentionally)
    "MD046",  # Code block style (indented blocks required inside MkDocs tabs)
    "MD051",  # Link fragments (cross-doc references resolved at MkDocs build time)
]
exclude = ["CHANGELOG.md", "site", "htmlcov", ".nox", ".venv"]

[tool.rumdl.per-file-ignores]
".github/PULL_REQUEST_TEMPLATE.md" = ["MD041"]  # PR template starts with checklist
".github/skills/**" = ["MD007"]  # Nested lists in skill files use variable indent
"README.md" = ["MD041"]  # README starts with logo, not heading
"docs/index.md" = ["MD041"]  # Landing page starts with logo, not heading
"CONTRIBUTING.md" = ["MD057"]  # Relative link to docs/ is valid at project level

[tool.ty]
# Pre-existing type issues in Azure ML SDK integration & Click parameter coercion.
# TODO: Gradually fix and re-enable these rules.
[tool.ty.rules]
invalid-assignment = "ignore"
invalid-argument-type = "ignore"
invalid-parameter-default = "ignore"
invalid-return-type = "ignore"
unresolved-attribute = "ignore"
unresolved-import = "ignore"
no-matching-overload = "ignore"
invalid-method-override = "ignore"

[tool.commitizen]
name = "cz_conventional_commits"
version_scheme = "semver"
version_provider = "scm"
update_changelog_on_bump = false
major_version_zero = true
tag_format = "v$version"
